refactor create_mosaic to class
