+++ cogeo_mosaic/backends/file.py
@@ -7,7 +7,7 @@ import functools
 
 import mercantile
 
-from cogeo_mosaic.model import MosaicJSON
+from cogeo_mosaic.mosaic import MosaicJSON
 from cogeo_mosaic.backends.base import BaseBackend
 from cogeo_mosaic.backends.utils import (
     _compress_gz_json,
