+++ cogeo_mosaic/backends/http.py
@@ -8,7 +8,7 @@ import functools
 import requests
 import mercantile
 
-from cogeo_mosaic.model import MosaicJSON
+from cogeo_mosaic.mosaic import MosaicJSON
 from cogeo_mosaic.backends.base import BaseBackend
 from cogeo_mosaic.backends.utils import _decompress_gz, get_assets_from_json
