--- cogeo_mosaic/model.py
@@ -1,26 +1,0 @@
-"""cogeo-mosaic models."""
-
-from typing import Dict, List, Optional, Tuple
-
-from pydantic import BaseModel, Field
-
-
-class MosaicJSON(BaseModel):
-    """
-    MosaicJSON model.
-
-    Based on https://github.com/developmentseed/mosaicjson-spec
-
-    """
-
-    mosaicjson: str
-    name: Optional[str]
-    description: Optional[str]
-    version: str = "1.0.0"
-    attribution: Optional[str]
-    minzoom: int = Field(0, ge=0, le=30)
-    maxzoom: int = Field(30, ge=0, le=30)
-    quadkey_zoom: Optional[int]
-    bounds: List[float] = Field([-180, -90, 180, 90])
-    center: Optional[Tuple[float, float, int]]
-    tiles: Dict[str, List[str]]
